refactor(lang): rename double type to float
【动机】
double 作为浮点类型名语义不直观（"双倍"看不出数字），float 更清晰。底层 BuiltinType::FLOAT 早已是这个类型，只是显示名叫 double。

【原理】
- 把类型名显示字符串 "double" 全部替换为 "float"（VM 注册、typeName、LSP、Tensor dtype 别名）。
- 底层类型、字节码、JIT 全不动（BuiltinType::FLOAT 不变）。
- 文档、测试、examples、插件（tmLanguage 类型列表 + documentation.json）同步。
- C++ 的 double 关键字、isDouble/asDouble 等内部方法名不改。

+++ src/frontend/Highlight.h
@@ -282,7 +282,7 @@ namespace jc {
     // ═══════════════════════════════════════════
     inline std::string colorizeType(const std::string& typeName) {
         if (!colorsEnabled) return "";
-        if (typeName == "double" || typeName == "BigInt" || typeName == "Fraction")
+        if (typeName == "float" || typeName == "BigInt" || typeName == "Fraction")
             return col(Ansi::BRIGHT_YELLOW);
         if (typeName == "Complex")
             return col(Ansi::BRIGHT_MAGENTA);
